Rename test-setup into test-helpers
